examples/copydatabase.cc: Fix to compile with GCC 2.95.
git-svn-id: file:///home/xapian-svn/xapian/trunk@8689 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/examples/copydatabase.cc
@@ -87,7 +87,7 @@ try {
 	}
 
 	// Calculate how many decimal digits there are in dbsize.
-	int width = static_cast<int>(log10(dbsize)) + 1;
+	int width = static_cast<int>(log10(double(dbsize))) + 1;
 
 	Xapian::doccount c = 0;
 	Xapian::PostingIterator it = db_in.postlist_begin("");
